requirements.txt
aiogram==3.7.0
aiohttp
feedparser
python-dotenv
flask